requirements.txt
click==8.1.3
Pillow==9.4.0
numpy==1.24.2